Rework for Modules task
