+++ Task7/package.json
@@ -6,9 +6,9 @@
     "start": "webpack-dev-server --hot --progress",
     "dev-serve": "http-server ./ -a 0.0.0.0 -p 9001",
     "prod-serve": "http-server ./ -a 0.0.0.0 -p 9000",
-    "ngc": "ngc -p src",
+    "ngc": "ngc -p tsconfig-aot.json",
     "dev": "set NODE_ENV=development&&webpack --config webpack.config.js && npm run dev-serve",
-    "build": "ngc && set NODE_ENV=production&&webpack -p --config prod.webpack.config.js && npm run prod-serve"
+    "build": "npm run ngc && set NODE_ENV=production&&webpack -p --config prod.webpack.config.js && npm run prod-serve"
   },
   "license": "MIT",
   "dependencies": {
@@ -31,16 +31,16 @@
     "@angular/compiler": "^2.2.4",
     "@angular/compiler-cli": "^2.4.4",
     "@angular/platform-server": "^2.4.4",
-    "@types/core-js": "^0.9.34",
+    "@types/core-js": "^0.9.35",
     "@types/googlemaps": "^3.25.36",
     "@types/node": "^6.0.45",
     "@types/whatwg-fetch": "0.0.32",
     "@types/whatwg-streams": "0.0.1",
     "angular2-template-loader": "^0.4.0",
-    "awesome-typescript-loader": "^3.0.0-beta.10",
+    "awesome-typescript-loader": "=3.0.0-beta.10",
     "codelyzer": "^2.0.0-beta.1",
     "css-loader": "^0.23.1",
-    "extract-text-webpack-plugin": "2.0.0-beta.4",
+    "extract-text-webpack-plugin": "=2.0.0-beta.4",
     "file-loader": "^0.8.5",
     "html-loader": "^0.4.3",
     "html-webpack-plugin": "^2.26.0",
@@ -50,9 +50,9 @@
     "rimraf": "^2.5.2",
     "style-loader": "^0.13.1",
     "to-string-loader": "^1.1.5",
-    "tslint": "^4.0.2",
-    "typescript": "^2.1.4",
-    "webpack": "2.2.0-rc.3",
+    "tslint": "=4.0.2",
+    "typescript": "=2.1.4",
+    "webpack": "=2.2.0-rc.3",
     "webpack-dev-server": "^1.14.1",
     "webpack-merge": "^0.14.0"
   }
