+++ Task7/src/app/app.module.ts
@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { MaterialModule } from '@angular/material';
 import { CoreModule } from './core/core.module';
-import { SharedModule } from './shared/shared.module';
 import { WeatherListModule } from './weatherList/index';
 import { MapModule } from './map/index';
 
@@ -13,8 +12,7 @@ import { AppComponent } from './app.component';
     imports: [
         BrowserModule, 
         MaterialModule.forRoot(), 
-        CoreModule,
-        SharedModule.forRoot(),
+        CoreModule.forRoot(),
         WeatherListModule, 
         MapModule
     ],
