+++ Task7/src/prod.app.ts
@@ -2,7 +2,7 @@ import 'zone.js';
 import 'reflect-metadata';
 
 import { platformBrowser } from '@angular/platform-browser';
-import { AppModuleNgFactory } from './app/app.module.ngfactory';
+import { AppModuleNgFactory } from '../aot/src/app/app.module.ngfactory';
 
 // Launch with the app module factory.
 platformBrowser().bootstrapModuleFactory(AppModuleNgFactory);
