+++ Task7/tsconfig-aot.json
@@ -1,0 +1,43 @@
+{
+    "compilerOptions": {
+        "target": "es5",
+        "module": "es2015",
+        "moduleResolution": "node",
+        "sourceMap": true,
+        "emitDecoratorMetadata": true,
+        "experimentalDecorators": true,
+        "lib": ["es5", "dom"],
+        "noImplicitAny": true,
+        "skipLibCheck": true,
+        "suppressImplicitAnyIndexErrors": true
+    },
+
+    "files": [
+        "src/app/app.module.ts",
+        "src/prod.app.ts"
+    ],
+
+    "typeRoots": [
+        "../node_modules/@types"
+    ],
+
+    "types": [
+        "core-js",
+        "whatwg-fetch",
+        "googlemaps"
+    ],
+
+    "exclude": [
+        "node_modules/*",
+        "dist",
+        "aot",
+        "111"
+    ],
+
+    "compileOnSave": false,
+
+    "angularCompilerOptions": {
+        "genDir": "aot",
+        "skipMetadataEmit" : true
+    }
+}
